chore: Update Slack "Add to Slack" button URL in slack.py
Update the URL of the Slack "Add to Slack" button in the slack.py file to include the client ID from the st.secrets. This change ensures that the button links to the correct authorization URL and enhances the user experience by providing a convenient way to connect the app with Slack.